Implemented addTiles method in Rack class
Implemented a method called addTiles in the Rack class, which adds an ArrayList of tiles back on to the player's rack. Will be used when the player removes tiles from the rack to place a word, but isn't able to do so. The addTiles method will enable the player to return the removed tiles back on to the players rack.