Added initial setter methods for PlayerCard
This commit contains initial setter methods for the PlayerCard class, including an updateScore method that updates the user's score when provided an integer value, and an updateRack method that updates the player's rack to display the new, updated letters when provided an array of chars. Also changed the ScrabbleFrame class so that the class now has a reference to all of the created PlayerCards.