Added update() method stub, and updated javadocs in ScrabbleGame class
Added a stub for the update() method, which will be responsible for processing an event and updating the GUI (view) according to this event. Also fixed an issue with the playerCards field not being initialized in the constructor, and updated the javadoc of the actionPerformed method to note that it should be removed once the Controller is implemented.