Added subscripts to Tile's text to display point values
Changed the string representation of a Tile to also display the Tile's point value as a subscript of the Tile letter. Also extended the width of the player panel, player card panel, and the main game panel to account for the extra width needed when adding subscripts (and large subscripts like 10) to the Tile's text. Also changed the appearance of the Tile to better resemble a real Scrabble tile (changed background to white, and foreground to black).