Changed appearance of rack tiles according to certain player actions
Changed the appearance of the center square on the Scrabble board. Changed the actionPerformed method in the ScrabbleController so that when the player clicks play, they are allowed to only select one tile at a time (which is depicted by the tile's background color changing to a light grey). Player also has the ability to select multiple tiles (multiple tiles will have a grey background color) to redraw, and the background color for those selected tiles will be reset to the default white color before going back into the tile bag. Changed the color of the submit button to red to notify the player that this action is important/must be clicked for their turn to be finalized. Also changed the font weight of the buttons to bold for better visibility.