Performed initial refactor of playWord() method
Performed initial refactor of playWord method where a new method called notifyInvalidPlacement() was created. This method takes in a message and creates a JOptionPane message dialog according to the message to notify the player of an invalid placement, and reverts the board to its initial state. This method is now invoked in the playWord method to reduce code duplication. Changes were also made to the checkWordBonus method, as the parameter for the method was misspelled, as well as a local variable within the method.