Implemented prompting user for valid number of AI players
Changed the configurePlayerInformation() method in the ScrabbleController class to now prompt the user for the number of AI players they would like to have in the Scrabble game, after the user has configured the regular players. The method will calculate the number of open spots based on the number of regular players initially specified, and will then verify if a valid number of AI players has been specified. AI players will only be available to be added if the number of regular players is less than 4. The user also has the option of having 0 AI players, regardless of the number of how many players they initially choose.
Also added public constants to the ScrabbleGameModel class, which are the minimum number of players (2) and maximum number of players (4). Initially, the configurePlayerInformation method was using hard-coded values to specify the maximum and minimum number of players. This was changed so that the method is now using the constants defined in the ScrabbleGameModel class (to make it easier to dynamically change the number of players for the game in the future).