Refactored configurePlayerInformation() method
Refactored the configurePlayerInformation() method by delegating some of the tasks it was performing to seperate methods. Since both regular and AI players need to be configured, configurePlayerInformation() only configures the regular players, and a new method called configureAIPlayerInformation() was created configure all the AI players. Next, the 'Let's play!' button is mapped to a new function called playScrabbleGame() which invokes both the configurePlayerInformation and configureAIPlayerInformation method and proceeds to start the game if both return no errors.
Also changed the logic for how AI players are selected. The user can now initially specify between 0 and 4 regular players, and must specify however many AI players to satisfy the rule of having at least 2 players and no more than 4 players in the Scrabble game. The acceptable range for AI players is also displayed in the message dialog.