Initial set up of AI player
Created a new interface called ScrabblePlayer which models a player in the Scrabble game. The PlayerModel now implements this interface, as well as a new class called AIPlayer (which will model an AI player in the Scrabble game).

Things to be done:
The AIPlayer class is currently implementing all the methods defined in the ScrabblePlayer interface. However, all of these methods in the AIPlayer class have not been implemented yet (empty bodies with default return statements). In addition, more methods will need to be added to the class to fully implement an AI player in the Scrabble game. Hence, the AIPlayer class currently serves as a starting point.