Added addScrabbleBoardLabels method
Added a new method in the ScrabbleFrameView class called addScrabbleBoardLabels, which contains the code implemented in the previous commit (adding horizontal and vertical labels to the Scrabble board), in order to reduce the size of the startGame method. Also added a field for the container JPanel which contains the board panel, the horizontal labelling panel, and the vertical labelling panel.