Revised/refactored addScrabbleBoardLabels() method
Revised and refactored the addScrabbleBoardLabels() method to eliminate bugs that were detected with the method. In the original version of the method, the horizontal and vertical 'labels' were each a single string, comprised of each of the different characters. Space between these characters were manually added to implement the board labelling. However, having the vertical and horizontal labels as strings was eventually determined to not be the best method to model the labels, as they would not be displayed as expected on different display sizes due to scaling. To eliminate this issue, the method was refactored to instead shift over to having dedicated JLabels for each character, rather than a single string to store all characters. To add spacing between the JLabels, the 'Box.createRigidArea' method was used, and was passed a Dimension object. By making this conversion, display scaling does not have an effect on the appearance of the vertical and horizontal labels.