Refactored font management in Scrabble game
Refactored the font management in the Scrabble game. Originally, a class called FontManager was created that loaded in two custom fonts into the GUI (FontManager objects were instanstiated whenever a particular class needed a font, as the class also had getter methods that would return the Font objects themselves). However, this would later serve to be an issue as it would significantly increase the compile time of the Scrabble game, and would also significantly increase the time taken to run tests. The main cause for this issue was the getResourceAsStream method and createFont method was called every time a certain class needed a font, which are typically heavy methods for compiling. To fix this issue, the FontManager class was no longer used and instead, a single method was written in the ScrabbleFrameView class called setupFonts(), which retrieves the local graphics environment, and registers the two fonts (which still calls the getResourceAsStream method) to the environment. This way, the Manrope font family (the font family of the two custom fonts being loaded in) is available to use for all classes, including ScrabbleFrameView, PlayerCardView, Square, and Tile. This implementation avoids the long compile time from each of the mentioned classes instantiating an instance of FontManager, which will essentially import and create the font each time. Furthermore, instead of doing the work multiple times, this new management of fonts only imports and registers the required fonts a single time.